backend/requirements/base.txt
# ==================================================
# JumushTap Backend — Base Requirements
# Shared across all environments
# ==================================================

# Django
Django>=5.0,<5.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3

# Database
psycopg2-binary>=2.9  # PostgreSQL adapter

# Cache & Queue
redis>=5.0
celery>=5.3
django-redis>=5.4

# Authentication & Security
django-phonenumber-field[phonenumbers]>=7.0
argon2-cffi>=23.1  # Secure password hashing
python-decouple>=3.8  # Environment variables

# API & Documentation
drf-spectacular>=0.27  # OpenAPI/Swagger

# Storage
django-storages>=1.14  # S3 support
boto3>=1.34  # AWS SDK
Pillow>=10.2  # Image processing

# Payments (choose based on PSP)
stripe>=7.0

# Firebase (Push notifications)
firebase-admin>=6.0

# HTTP & Networking
requests>=2.31
urllib3>=2.1

# Utils
python-dateutil>=2.8
pytz>=2024.1

# CORS
django-cors-headers>=4.3

# Logging
python-json-logger>=2.0  # Structured logs
